# src/openapi/api.yml
openapi: '3.0.2'
info:
  title: API Title
  version: '1.0'
servers:
  - url: http://localhost:3610/v1
...
paths:
...
  /accounts/{accountId}:
    get:
      tags:
        - Accounts 📝
      security:
        - bearerAuth: []
      summary: Retrieve a specific account by ID
      description: |
        Retrieve a specific account by its unique ID.
        This endpoint allows you to fetch details of a specific account using its unique identifier (UUIDv4).
      x-eov-operation-handler: accountController
      operationId: findById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Account'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - Accounts 📝
      security:
        - bearerAuth: []
      summary: Delete a specific account by ID
      description: |
        Delete a specific account by its unique ID.
        This endpoint allows you to remove details of a specific account using its unique identifier (UUIDv4).
      x-eov-operation-handler: accountController
      operationId: deleteById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                required:
                  - statusCode
                  - message
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /accounts/{accountId}/applications:
    post:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Create a new application
      description: |
        Adds a new application entry to the system.
        Use this endpoint to create a new application along with its details.
      x-eov-operation-handler: applicationController
      operationId: create
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the application.
                  example: My App
                description:
                  type: string
                  description: A brief description of the application.
                  example: This is a fantastic application
                image:
                  type: string
                  description: URL or path to an image representing the application.
                  nullable: true
                enabled:
                  type: boolean
                  description: A flag indicating whether the application is enabled.
                  example: true
                  default: true
              required:
                - name
                - enabled
              additionalProperties: false
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 201
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: Created
                  data:
                    $ref: '#/components/schemas/Application'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    get:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Retrieve a list of applications
      description: This endpoint allows retrieving a list of users from a specific account.
      x-eov-operation-handler: applicationController
      operationId: findAndCountAll
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - name: page
          in: query
          description: The page number for pagination.
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: limit
          in: query
          description: The limit of results per page.
          required: false
          schema:
            type: integer
            format: int64
            default: 20
        - name: find
          in: query
          description: A search term to filter results.
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: >
            Order of the results: 
              
              - `a-z` - Order by name ascending. From A to Z
              - `z-a` - Order by name descending. From Z to A
          required: false
          schema:
            type: string
            enum:
              - a-z
              - z-a
            default: a-z
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  count:
                    type: integer
                    description: Total count of the results.
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Application'
                required:
                  - statusCode
                  - message
                  - count
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /accounts/{accountId}/applications/{applicationId}:
    get:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Retrieve a specific application by ID
      description: |
        Retrieve a specific application by its unique ID.
        This endpoint allows you to fetch details of a specific application using its unique identifier (UUIDv4).
      x-eov-operation-handler: applicationController
      operationId: findById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Update a specific application by ID
      description: |
        Update a specific application by its unique ID.
        This endpoint allows you to update details of a specific application using its unique identifier (UUIDv4).
      x-eov-operation-handler: applicationController
      operationId: updateById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the application.
                  example: My App
                description:
                  type: string
                  description: A brief description of the application.
                  example: This is a fantastic application
                enabled:
                  type: boolean
                  description: A flag indicating whether the application is enabled.
                  example: true
                  default: true
              additionalProperties: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  data:
                    $ref: '#/components/schemas/Application'
                required:
                  - statusCode
                  - message
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - Applications 📱
      security:
        - bearerAuth: []
      summary: Delete a specific application by ID
      description: |
        Delete a specific application by its unique ID.
        This endpoint allows you to delete details of a specific application using its unique identifier (UUIDv4).
      x-eov-operation-handler: applicationController
      operationId: deleteById
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: path
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                required:
                  - statusCode
                  - message
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /accounts/{accountId}/staff:
    get:
      tags:
        - Staff 👨‍🏫
      security:
        - bearerAuth: []
      summary: Retrieve a list of staff members
      description: This endpoint allows retrieving a list of staff members from a specific account.
      x-eov-operation-handler: staffController
      operationId: findAndCountAll
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
            description: The unique identifier for the account (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
          required: true
        - in: query
          name: applicationId
          schema:
            type: string
            description: The unique identifier for the application (UUIDv4).
            example: '123e4567-e89b-12d3-a456-426614174000'
        - name: page
          in: query
          description: The page number for pagination.
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: limit
          in: query
          description: The limit of results per page.
          required: false
          schema:
            type: integer
            format: int64
            default: 20
        - name: find
          in: query
          description: A search term to filter results.
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: >
            Order of the results: 
              
              - `a-z` - Order by name ascending. From A to Z
              - `z-a` - Order by name descending. From Z to A
          required: false
          schema:
            type: string
            enum:
              - a-z
              - z-a
            default: a-z
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: HTTP status code indicating success.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the success.
                    default: OK
                  count:
                    type: integer
                    description: Total count of the results.
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Staff'
                required:
                  - statusCode
                  - message
                  - count
                  - data
                additionalProperties: false
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
  schemas:
    BadRequest:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a bad request error.
          default: 400
        message:
          type: string
          description: Descriptive message of the error.
          default: Bad Request
        errorCode:
          type: string
          enum:
            - BAD_REQUEST
          description: >
            Possible error codes:
              
            - `BAD_REQUEST`: Indicates a bad request error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    Unauthorized:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating an unauthorized error.
          default: 401
        message:
          type: string
          description: Descriptive message of the error.
          default: Unauthorized
        errorCode:
          type: string
          enum:
            - UNAUTHORIZED
          description: >
            Possible error codes:
              
            - `UNAUTHORIZED`: Indicates an unauthorized error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    Forbidden:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a forbidden error.
          default: 403
        message:
          type: string
          description: Descriptive message of the error.
          default: Forbidden
        errorCode:
          type: string
          enum:
            - FORBIDDEN
          description: >
            Possible error codes:
              
            - `FORBIDDEN`: Indicates a forbidden error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    NotFound:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a not found error.
          default: 404
        message:
          type: string
          description: Descriptive message of the error.
          default: Not Found
        errorCode:
          type: string
          enum:
            - NOT_FOUND
          description: >
            Possible error codes:
              
            - `NOT_FOUND`: Indicates a not found error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    Conflict:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating a conflict error.
          default: 409
        message:
          type: string
          description: Descriptive message of the error.
          default: Conflict
        errorCode:
          type: string
          enum:
            - CONFLICT
          description: >
            Possible error codes:
              
            - `CONFLICT`: Indicates a conflict error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: integer
          format: int64
          description: HTTP status code indicating an error.
          default: 500
        message:
          type: string
          description: Descriptive message of the error.
          default: Internal Server Error
        errorCode:
          type: string
          enum:
            - INTERNAL_SERVER_ERROR
          description: >
            Possible error codes:
              
            - `INTERNAL_SERVER_ERROR`: Indicates an error.'
      required:
        - statusCode
        - message
        - errorCode
      additionalProperties: false
...
    Account:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the account (UUIDv4).
          example: '123e4567-e89b-12d3-a456-426614174000'
        createdAt:
          type: string
          format: date-time
          description: The time when the account was created.
          example: '2023-05-18T12:34:56Z'
        updatedAt:
          type: string
          format: date-time
          description: The time when the account was last updated.
          example: '2023-05-18T12:34:56Z'
        enabled:
          type: boolean
          description: A flag indicating whether the account is enabled.
          example: true
      required:
        - id
        - createdAt
        - updatedAt
        - enabled
      additionalProperties: false
    Staff:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the account (UUIDv4).
          example: '123e4567-e89b-12d3-a456-426614174000'
        createdAt:
          type: string
          format: date-time
          description: The time when the account was created.
          example: '2023-05-18T12:34:56Z'
        updatedAt:
          type: string
          format: date-time
          description: The time when the account was last updated.
          example: '2023-05-18T12:34:56Z'
        accountId:
          type: string
          description: The unique identifier for the account (UUIDv4).
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          description: The name of the user.
          example: John Doe
        email:
          type: string
          description: The email address of the user.
          example: john@example.com
        type:
          type: string
          description: The type of user
          example: ACCOUNT_OWNER
          enum:
            - ACCOUNT_OWNER
        image:
          type: string
          description: URL or path to an image representing the tournament.
          nullable: true
        enabled:
          type: boolean
          description: A flag indicating whether the account is enabled.
          example: true
      required:
        - id
        - createdAt
        - updatedAt
        - accountId
        - name
        - email
        - type
        - image
        - enabled
      additionalProperties: false
    Application:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the application (UUIDv4).
          example: '123e4567-e89b-12d3-a456-426614174000'
        createdAt:
          type: string
          format: date-time
          description: The time when the application was created.
          example: '2023-05-18T12:34:56Z'
        updatedAt:
          type: string
          format: date-time
          description: The time when the application was last updated.
          example: '2023-05-18T12:34:56Z'
        accountId:
          type: string
          description: The unique identifier for the account (UUIDv4).
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          description: The name of the application.
          example: My App
        description:
          type: string
          description: A brief description of the application.
          example: This is a fantastic application
          nullable: true
        image:
          type: string
          description: The image logo of the application.
          nullable: true
        enabled:
          type: boolean
          description: A flag indicating whether the application is enabled.
          example: true
      required:
        - id
        - createdAt
        - updatedAt
        - accountId
        - name
        - description
        - image
        - enabled
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT